requirements.txt
PyPDF2==3.0.1
sentence-transformers==2.2.2
pinecone-client==3.0.0
numpy==1.24.3
torch==2.0.1
transformers==4.30.2